+++ Source/SwiftLintBuiltInRules/Rules/Lint/ProhibitedSuperRule.swift
@@ -80,7 +80,7 @@ private extension ProhibitedSuperRule {
             guard let body = node.body,
                   node.modifiers.contains(keyword: .override),
                   !node.modifiers.containsStaticOrClass,
-                  case let name = node.resolvedName(),
+                  case let name = node.resolvedName,
                   configuration.resolvedMethodNames.contains(name),
                   node.numberOfCallsToSuper() > 0 else {
                 return