+++ Source/SwiftLintBuiltInRules/Rules/Style/IdentifierNameRule.swift
@@ -1,7 +1,9 @@
 import Foundation
-import SourceKittenFramework
+import SwiftLintCore
+import SwiftSyntax
 
-struct IdentifierNameRule: ASTRule {
+@SwiftSyntaxRule
+struct IdentifierNameRule: Rule {
     var configuration = NameConfiguration<Self>(minLengthWarning: 3,
                                                 minLengthError: 2,
                                                 maxLengthWarning: 40,
@@ -21,102 +23,203 @@ struct IdentifierNameRule: ASTRule {
         triggeringExamples: IdentifierNameRuleExamples.triggeringExamples,
         deprecatedAliases: ["variable_name"]
     )
+}
 
-    func validate(
-        file: SwiftLintFile,
-        kind: SwiftDeclarationKind,
-        dictionary: SourceKittenDictionary
-    ) -> [StyleViolation] {
-        guard !dictionary.enclosedSwiftAttributes.contains(.override) else {
-            return []
+private extension IdentifierNameRule {
+    final class Visitor: ViolationsSyntaxVisitor<ConfigurationType> {
+        override func visitPost(_ node: AccessorParametersSyntax) {
+            collectViolations(from: .variable(name: node.name.text, isStatic: false, isPrivate: false), on: node.name)
         }
 
-        return validateName(dictionary: dictionary, kind: kind).map { name, offset in
-            guard let firstCharacter = name.first, !configuration.shouldExclude(name: name) else {
-                return []
+        override func visitPost(_ node: ClosureParameterSyntax) {
+            let name = (node.secondName ?? node.firstName).text.leadingDollarStripped
+            if node.modifiers.contains(keyword: .override) {
+                return
             }
-            let type = self.type(for: kind)
-            if !SwiftDeclarationKind.functionKinds.contains(kind) {
-                if !configuration.allowedSymbolsAndAlphanumerics.isSuperset(of: CharacterSet(charactersIn: name)) {
-                    return [
-                        StyleViolation(ruleDescription: Self.description,
-                                       severity: configuration.unallowedSymbolsSeverity.severity,
-                                       location: Location(file: file, byteOffset: offset),
-                                       reason: """
-                                            \(type) name '\(name)' should only contain alphanumeric and other \
-                                            allowed characters
-                                            """)
-                    ]
+            collectViolations(from: .variable(name: name, isStatic: false, isPrivate: false), on: node.firstName)
+        }
+
+        override func visitPost(_ node: ClosureShorthandParameterSyntax) {
+            collectViolations(
+                from: .variable(name: node.name.text.leadingDollarStripped, isStatic: false, isPrivate: false),
+                on: node.name
+            )
+        }
+
+        override func visitPost(_ node: EnumCaseElementSyntax) {
+            collectViolations(from: .enumElement(name: node.name.text), on: node.name)
+        }
+
+        override func visitPost(_ node: EnumCaseParameterSyntax) {
+            if let param = node.secondName ?? node.firstName, let position = node.firstName {
+                collectViolations(from: .variable(name: param.text, isStatic: false, isPrivate: false), on: position)
+            }
+        }
+
+        override func visitPost(_ node: FunctionDeclSyntax) {
+            let name = node.name.text
+            if node.modifiers.contains(keyword: .override) || name.isOperator {
+                return
+            }
+            let type = NamedDeclType.function(
+                name: name,
+                resolvedName: node.resolvedName,
+                isPrivate: node.modifiers.containsPrivateOrFileprivate()
+            )
+            let staticKeyword = node.modifiers.staticOrClassModifier
+            collectViolations(from: type, on: staticKeyword?.name ?? node.funcKeyword)
+        }
+
+        override func visitPost(_ node: FunctionParameterSyntax) {
+            if !node.modifiers.contains(keyword: .override) {
+                let name = (node.secondName ?? node.firstName).text
+                collectViolations(from: .variable(name: name, isStatic: false, isPrivate: false), on: node.firstName)
+            }
+        }
+
+        override func visitPost(_ node: IdentifierPatternSyntax) {
+            let varDecl = node.enclosingVarDecl
+            if varDecl?.modifiers.contains(keyword: .override) ?? false {
+                return
+            }
+            let type = NamedDeclType.variable(
+                name: node.identifier.text,
+                isStatic: varDecl?.modifiers.contains(keyword: .static) ?? false,
+                isPrivate: varDecl?.modifiers.containsPrivateOrFileprivate() ?? false
+            )
+            let staticKeyword = varDecl?.modifiers.staticOrClassModifier
+            let position = staticKeyword?.name ?? varDecl?.bindingSpecifier ?? node.identifier
+            collectViolations(from: type, on: position)
+        }
+
+        private func collectViolations(from type: NamedDeclType, on token: TokenSyntax) {
+            if let violation = violates(type) {
+                violations.append(ReasonedRuleViolation(
+                    position: token.positionAfterSkippingLeadingTrivia,
+                    reason: violation.reason,
+                    severity: violation.severity
+                ))
+            }
+        }
+
+        private func violates(_ type: NamedDeclType) -> (reason: String, severity: ViolationSeverity)? {
+            guard !configuration.shouldExclude(name: type.name), type.name != "_",
+                  let firstCharacter = type.name.first else {
+                return nil
+            }
+            if case .function = type {
+                // Do not perform additional checks.
+            } else {
+                if !configuration.allowedSymbolsAndAlphanumerics.isSuperset(of: CharacterSet(charactersIn: type.name)) {
+                    let reason = """
+                        \(type) should only contain alphanumeric and other \
+                        allowed characters
+                        """
+                    return (reason, configuration.unallowedSymbolsSeverity.severity)
                 }
 
-                if let severity = configuration.severity(forLength: name.count) {
-                    let reason = "\(type) name '\(name)' should be between " +
-                        "\(configuration.minLengthThreshold) and " +
-                        "\(configuration.maxLengthThreshold) characters long"
-                    return [
-                        StyleViolation(ruleDescription: Self.description,
-                                       severity: severity,
-                                       location: Location(file: file, byteOffset: offset),
-                                       reason: reason)
-                    ]
+                if let severity = configuration.severity(forLength: type.name.count) {
+                    let reason = """
+                        \(type) should be between \
+                        \(configuration.minLengthThreshold) and \
+                        \(configuration.maxLengthThreshold) characters long
+                        """
+                    return (reason, severity)
                 }
             }
-
             if configuration.allowedSymbols.contains(String(firstCharacter)) {
-                return []
+                return nil
             }
             if let caseCheckSeverity = configuration.validatesStartWithLowercase.severity,
-                kind != .varStatic && name.isViolatingCase && !name.isOperator {
-                let reason = "\(type) name '\(name)' should start with a lowercase character"
-                return [
-                    StyleViolation(ruleDescription: Self.description,
-                                   severity: caseCheckSeverity,
-                                   location: Location(file: file, byteOffset: offset),
-                                   reason: reason)
-                ]
+               !type.isStatic, type.name.isViolatingCase, !type.name.isOperator {
+                let reason = "\(type) should start with a lowercase character"
+                return (reason, caseCheckSeverity)
             }
+            return nil
+        }
+    }
+}
 
-            return []
-        } ?? []
+private extension DeclModifierListSyntax {
+    var staticOrClassModifier: DeclModifierSyntax? {
+        first { ["static", "class"].contains($0.name.text) }
+    }
+}
+
+private extension IdentifierPatternSyntax {
+    var enclosingVarDecl: VariableDeclSyntax? {
+        let identifierDecl =
+             parent?.as(PatternBindingSyntax.self)?
+            .parent?.as(PatternBindingListSyntax.self)?
+            .parent?.as(VariableDeclSyntax.self)
+        if identifierDecl != nil {
+            return identifierDecl
+        }
+        return
+             parent?.as(TuplePatternElementSyntax.self)?
+            .parent?.as(TuplePatternElementListSyntax.self)?
+            .parent?.as(TuplePatternSyntax.self)?
+            .parent?.as(PatternBindingSyntax.self)?
+            .parent?.as(PatternBindingListSyntax.self)?
+            .parent?.as(VariableDeclSyntax.self)
+    }
+}
+
+private extension VariableDeclSyntax {
+    var allDeclaredNames: [String] {
+        bindings
+            .map(\.pattern)
+            .flatMap { pattern -> [String] in
+                if let id = pattern.as(IdentifierPatternSyntax.self) {
+                    [id.identifier.text]
+                } else if let tuple = pattern.as(TuplePatternSyntax.self) {
+                    tuple.elements.compactMap {
+                        $0.pattern.as(IdentifierPatternSyntax.self)?.identifier.text
+                    }
+                } else {
+                    []
+                }
+            }
+    }
+}
+
+private enum NamedDeclType: CustomStringConvertible {
+    case function(name: String, resolvedName: String, isPrivate: Bool)
+    case enumElement(name: String)
+    case variable(name: String, isStatic: Bool, isPrivate: Bool)
+
+    var description: String {
+        switch self {
+        case let .function(_, resolvedName, _): "Function name '\(resolvedName)'"
+        case let .enumElement(name): "Enum element name '\(name)'"
+        case let .variable(name, _, _): "Variable name '\(name)'"
+        }
     }
 
-    private func validateName(
-        dictionary: SourceKittenDictionary,
-        kind: SwiftDeclarationKind
-    ) -> (name: String, offset: ByteCount)? {
-        guard
-            var name = dictionary.name,
-            let offset = dictionary.offset,
-            kinds.contains(kind),
-            !name.hasPrefix("$")
-        else { return nil }
-
-        if
-            kind == .enumelement,
-            let parenIndex = name.firstIndex(of: "("),
-            parenIndex > name.startIndex
-        {
-            let index = name.index(before: parenIndex)
-            name = String(name[...index])
+    var isStatic: Bool {
+        switch self {
+        case .function, .enumElement: false
+        case let .variable(_, isStatic, _): isStatic
         }
-
-        return (name.nameStrippingLeadingUnderscoreIfPrivate(dictionary), offset)
     }
 
-    private let kinds: Set<SwiftDeclarationKind> = {
-        return SwiftDeclarationKind.variableKinds
-            .union(SwiftDeclarationKind.functionKinds)
-            .union([.enumelement])
-    }()
-
-    private func type(for kind: SwiftDeclarationKind) -> String {
-        if SwiftDeclarationKind.functionKinds.contains(kind) {
-            return "Function"
-        } else if kind == .enumelement {
-            return "Enum element"
-        } else {
-            return "Variable"
+    var isPrivate: Bool {
+        switch self {
+        case let .function(_, _, isPrivate): isPrivate
+        case .enumElement: false
+        case let .variable(_, _, isPrivate): isPrivate
         }
+    }
+
+    var name: String {
+        let name = switch self {
+        case let .function(name, _, _): name
+        case let .enumElement(name): name
+        case let .variable(name, _, _): name
+        }
+        return name
+            .trimmingCharacters(in: CharacterSet(charactersIn: "`"))
+            .strippingLeadingUnderscore(if: isPrivate)
     }
 }
 
@@ -138,11 +241,15 @@ private extension String {
         return operators.contains(where: hasPrefix)
     }
 
-    func nameStrippingLeadingUnderscoreIfPrivate(_ dict: SourceKittenDictionary) -> String {
-        if let acl = dict.accessibility,
-            acl.isPrivate && first == "_" {
-            return String(self[index(after: startIndex)...])
-        }
-        return self
+    func strippingLeadingUnderscore(if isPrivate: Bool) -> Self {
+        isPrivate && first == "_" ? allButFirstCharacter : self
+    }
+
+    var leadingDollarStripped: Self {
+        first == "$" ? allButFirstCharacter : self
+    }
+
+    private var allButFirstCharacter: String {
+        String(self[index(after: startIndex)...])
     }
 }